refactor(arch): Phase 5 — PR #486 리뷰 수정
- B1: proxy.test.ts jest.mock → 배럴 경로
- B2/B3: useCurrentUser/useModelGate → entities/session/actions 배럴
- B4: BCRYPT_DEFAULT_SALT_ROUNDS 배럴 노출 제거
- S1: eslint features cross-import 주석에 Phase 7 해소 계획 추가
