+++ src/__tests__/proxy.test.ts
@@ -1,9 +1,7 @@
-// proxy.ts는 `@/entities/session/lib/sessionCookie`에서 AUTH_SESSION_COOKIE_NAME을
-// import한다 (siglens-core가 아님). 잘못된 모듈을 mock하면 mock이 실제로 적용되지 않고
-// 두 source가 우연히 같은 literal('siglens_session')을 공유한 덕에 통과해온 상태였다.
+// proxy.ts는 `@/entities/session` 배럴에서 AUTH_SESSION_COOKIE_NAME을 import한다.
 // sessionCookie.ts는 외부 의존 없는 순수 상수 모듈이라 실제로 mock할 필요도 없지만,
-// 명시적으로 의도를 드러내기 위해 올바른 모듈 경로로 mock을 유지한다.
-jest.mock('@/entities/session/lib/sessionCookie', () => ({
+// 명시적으로 의도를 드러내기 위해 배럴 경로로 mock을 유지한다.
+jest.mock('@/entities/session', () => ({
     AUTH_SESSION_COOKIE_NAME: 'siglens_session',
 }));
 jest.mock('next/server', () => ({
